uni-quiz-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Telegram
python-telegram-bot==20.7

# File handling
python-multipart==0.0.6
aiofiles==23.2.1

# Excel export
openpyxl==3.1.2

# Validation & Utils
email-validator==2.1.0.post1
httpx==0.25.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3


# Development
black==23.12.1
isort==5.13.2
